CEF控件子窗口模式：拖出标签和拖入标签抓图功能的代码结构调整，统一放到平台相关的文件中。
