+++ duilib/CEFControl/CefControlNative.cpp
@@ -5,11 +5,6 @@
 #include "duilib/Core/Box.h"
 #include "duilib/Core/GlobalManager.h"
 
-#ifdef DUILIB_BUILD_FOR_LINUX
-    #include <X11/Xlib.h>
-    #include <X11/Xutil.h>
-#endif
-
 namespace ui {
 
 CefControlNative::CefControlNative(ui::Window* pWindow):
@@ -212,14 +207,7 @@ std::shared_ptr<IBitmap> CefControlNative::MakeImageSnapshot()
     std::vector<uint8_t> bitmap;
     int32_t width = 0;
     int32_t height = 0;
-    bool bRet = false;
-#ifdef DUILIB_BUILD_FOR_WIN
-    bRet = CaptureWindowBitmap_Win32(bitmap, width, height);
-#elif defined (DUILIB_BUILD_FOR_MACOS)
-    bRet = CaptureWindowBitmap_Mac(bitmap, width, height);
-#elif defined (DUILIB_BUILD_FOR_LINUX) || defined (DUILIB_BUILD_FOR_FREEBSD)
-    bRet = CaptureWindowBitmap_X11(bitmap, width, height);
-#endif
+    bool bRet = CaptureCefWindowBitmap(GetCefWindowHandle(), bitmap, width, height);
     if (bRet && (width > 0) && (height > 0) && ((int32_t)bitmap.size() == (width * height * 4))) {
         std::shared_ptr<IBitmap> spBitmap;
         IRenderFactory* pRenderFactory = GlobalManager::Instance().GetRenderFactory();
@@ -236,176 +224,5 @@ std::shared_ptr<IBitmap> CefControlNative::MakeImageSnapshot()
     }
     return nullptr;
 }
-
-#if defined (DUILIB_BUILD_FOR_WIN)
-bool CefControlNative::CaptureWindowBitmap_Win32(std::vector<uint8_t>& bitmap, int32_t& width, int32_t& height)
-{
-    HWND hwnd = GetCefWindowHandle();
-    if (!::IsWindow(hwnd)) {
-        return false;
-    }
-    // 获取窗口尺寸
-    RECT rect = { 0, 0, 0, 0 };
-    if (!GetClientRect(hwnd, &rect)) {
-        return false;
-    }
-
-    width = rect.right - rect.left;
-    height = rect.bottom - rect.top;
-
-    if (width <= 0 || height <= 0) {
-        return false;
-    }
-
-    // 创建设备上下文
-    HDC hdcScreen = ::GetDC(nullptr);
-    if (hdcScreen == nullptr) {
-        return false;
-    }
-    HDC hdcWindow = ::GetDC(hwnd);
-    if (hdcWindow == nullptr) {
-        ::ReleaseDC(nullptr, hdcScreen);
-        return false;
-    }
-
-    HDC hdcMemDC = ::CreateCompatibleDC(hdcWindow);
-    if (hdcMemDC == nullptr) {
-        ::ReleaseDC(nullptr, hdcScreen);
-        ::ReleaseDC(hwnd, hdcWindow);
-        return false;
-    }
-
-    // 创建位图
-    HBITMAP hBitmap = ::CreateCompatibleBitmap(hdcWindow, width, height);
-    if (hBitmap == nullptr) {
-        ::DeleteDC(hdcMemDC);
-        ::ReleaseDC(nullptr, hdcScreen);
-        ::ReleaseDC(hwnd, hdcWindow);
-        return false;
-    }
-
-    HGDIOBJ hOldObj = ::SelectObject(hdcMemDC, hBitmap);
-
-    // 拷贝屏幕内容到位图
-    ::BitBlt(hdcMemDC, 0, 0, width, height, hdcWindow, 0, 0, SRCCOPY);
-
-    // 获取位图信息
-    BITMAPINFOHEADER bi;
-    bi.biSize = sizeof(BITMAPINFOHEADER);
-    bi.biWidth = width;
-    bi.biHeight = -height;  // 正数表示从下到上，负数表示从上到下
-    bi.biPlanes = 1;
-    bi.biBitCount = 32;
-    bi.biCompression = BI_RGB;
-    bi.biSizeImage = 0;
-    bi.biXPelsPerMeter = 0;
-    bi.biYPelsPerMeter = 0;
-    bi.biClrUsed = 0;
-    bi.biClrImportant = 0;
-
-    // 分配内存并获取位图数据
-    bitmap.resize(width * height * 4);
-    ::GetDIBits(hdcMemDC, hBitmap, 0, height, bitmap.data(), (BITMAPINFO*)&bi, DIB_RGB_COLORS);
-
-    // 清理资源
-    ::SelectObject(hdcMemDC, hOldObj);
-    ::DeleteObject(hBitmap);
-    ::DeleteDC(hdcMemDC);
-    ::ReleaseDC(nullptr, hdcScreen);
-    ::ReleaseDC(hwnd, hdcWindow);
-
-    return true;
-}
-#elif defined (DUILIB_BUILD_FOR_MACOS)
-bool CefControlNative::CaptureWindowBitmap_Mac(std::vector<uint8_t>& bitmap, int32_t& width, int32_t& height)
-{
-    return false;
-}
-#elif defined (DUILIB_BUILD_FOR_LINUX) || defined (DUILIB_BUILD_FOR_FREEBSD)
-bool CefControlNative::CaptureWindowBitmap_X11(std::vector<uint8_t>& bitmap, int32_t& width, int32_t& height)
-{
-    // 检查X11环境更加健壮
-    const char* sessionType = std::getenv("XDG_SESSION_TYPE");
-    if (!sessionType || (std::string(sessionType) != "x11" && std::string(sessionType) != "X11")) {
-        // 尝试使用DISPLAY环境变量进行二次检查
-        const char* displayEnv = std::getenv("DISPLAY");
-        if (!displayEnv || !*displayEnv) {
-            return false;
-        }
-    }
-
-    Display* display = ::XOpenDisplay(nullptr);
-    if (!display) {
-        return false;
-    }
-
-    // RAII资源管理
-    struct DisplayCloser {
-        Display* d;
-        ~DisplayCloser() { if(d) ::XCloseDisplay(d); }
-    } closer{display};
-
-    ::Window x11Window = GetCefWindowHandle();
-
-    // 获取窗口尺寸
-    ::XWindowAttributes gwa;
-    if (!::XGetWindowAttributes(display, x11Window, &gwa)) {
-        return false;
-    }
-    width = gwa.width;
-    height = gwa.height;
-    if (width <= 0 || height <= 0) {
-        return false;
-    }
-
-    // 获取窗口内容
-    XImage* image = XGetImage(display, x11Window, 0, 0, width, height, AllPlanes, ZPixmap);
-    if (!image) {
-        return false;
-    }
-
-    // RAII管理XImage资源
-    struct ImageDestroyer {
-        XImage* img;
-        ~ImageDestroyer() { if(img) XDestroyImage(img); }
-    } imgDestroyer{image};
-
-    // 分配内存并复制像素数据
-    bitmap.resize(width * height * 4);
-
-    // 使用更安全的像素格式转换
-    bool isRgbOrder = (image->red_mask == 0xFF0000);
-    bool isBgrOrder = (image->blue_mask == 0xFF0000);
-    
-    // 使用XGetPixel作为安全的像素获取方式
-    for (int y = 0; y < height; y++) {
-        for (int x = 0; x < width; x++) {
-            unsigned long pixel = XGetPixel(image, x, y);
-            unsigned char r, g, b;
-            
-            if (isRgbOrder) {
-                r = (pixel >> 16) & 0xFF;
-                g = (pixel >> 8) & 0xFF;
-                b = pixel & 0xFF;
-            } else if (isBgrOrder) {
-                r = pixel & 0xFF;
-                g = (pixel >> 8) & 0xFF;
-                b = (pixel >> 16) & 0xFF;
-            } else {
-                // 无法确定顺序，使用灰度
-                r = g = b = (pixel * 255) / ((1 << image->bits_per_pixel) - 1);
-            }
-            
-            int index = (y * width + x) * 4;
-            bitmap[index] = r;
-            bitmap[index + 1] = g;
-            bitmap[index + 2] = b;
-            bitmap[index + 3] = 255;
-        }
-    }
-
-    return true;
-}
-#endif
 
 }
