+++ duilib/CEFControl/CefWindowUtils.h
@@ -2,6 +2,7 @@
 #define UI_CEF_CONTROL_CEF_WINDOW_UTILS_H_
 
 #include "CefControl.h"
+#include <vector>
 
 namespace ui
 {
@@ -16,6 +17,10 @@ void SetCefWindowVisible(CefWindowHandle cefWindow, CefControl* pCefControl);
 /** 设置CEF关联的窗口的父窗口(父窗口为pCefControl的关联窗口)
 */
 void SetCefWindowParent(CefWindowHandle cefWindow, CefControl* pCefControl);
+
+/** 抓取CEF窗口的截图为位图数据
+*/
+bool CaptureCefWindowBitmap(CefWindowHandle cefWindow, std::vector<uint8_t>& bitmap, int32_t& width, int32_t& height);
 
 } //namespace ui
