+++ duilib/CEFControl/CefWindowUtils_Windows.cpp
@@ -64,6 +64,85 @@ void SetCefWindowParent(CefWindowHandle cefWindow, CefControl* pCefControl)
     ::SetWindowLong(hParent, GWL_STYLE, style | WS_CLIPSIBLINGS | WS_CLIPCHILDREN);
 }
 
+bool CaptureCefWindowBitmap(CefWindowHandle cefWindow, std::vector<uint8_t>& bitmap, int32_t& width, int32_t& height)
+{
+    HWND hwnd = cefWindow;
+    if (!::IsWindow(hwnd)) {
+        return false;
+    }
+    // 获取窗口尺寸
+    RECT rect = { 0, 0, 0, 0 };
+    if (!GetClientRect(hwnd, &rect)) {
+        return false;
+    }
+
+    width = rect.right - rect.left;
+    height = rect.bottom - rect.top;
+
+    if (width <= 0 || height <= 0) {
+        return false;
+    }
+
+    // 创建设备上下文
+    HDC hdcScreen = ::GetDC(nullptr);
+    if (hdcScreen == nullptr) {
+        return false;
+    }
+    HDC hdcWindow = ::GetDC(hwnd);
+    if (hdcWindow == nullptr) {
+        ::ReleaseDC(nullptr, hdcScreen);
+        return false;
+    }
+
+    HDC hdcMemDC = ::CreateCompatibleDC(hdcWindow);
+    if (hdcMemDC == nullptr) {
+        ::ReleaseDC(nullptr, hdcScreen);
+        ::ReleaseDC(hwnd, hdcWindow);
+        return false;
+    }
+
+    // 创建位图
+    HBITMAP hBitmap = ::CreateCompatibleBitmap(hdcWindow, width, height);
+    if (hBitmap == nullptr) {
+        ::DeleteDC(hdcMemDC);
+        ::ReleaseDC(nullptr, hdcScreen);
+        ::ReleaseDC(hwnd, hdcWindow);
+        return false;
+    }
+
+    HGDIOBJ hOldObj = ::SelectObject(hdcMemDC, hBitmap);
+
+    // 拷贝屏幕内容到位图
+    ::BitBlt(hdcMemDC, 0, 0, width, height, hdcWindow, 0, 0, SRCCOPY);
+
+    // 获取位图信息
+    BITMAPINFOHEADER bi;
+    bi.biSize = sizeof(BITMAPINFOHEADER);
+    bi.biWidth = width;
+    bi.biHeight = -height;  // 正数表示从下到上，负数表示从上到下
+    bi.biPlanes = 1;
+    bi.biBitCount = 32;
+    bi.biCompression = BI_RGB;
+    bi.biSizeImage = 0;
+    bi.biXPelsPerMeter = 0;
+    bi.biYPelsPerMeter = 0;
+    bi.biClrUsed = 0;
+    bi.biClrImportant = 0;
+
+    // 分配内存并获取位图数据
+    bitmap.resize(width * height * 4);
+    ::GetDIBits(hdcMemDC, hBitmap, 0, height, bitmap.data(), (BITMAPINFO*)&bi, DIB_RGB_COLORS);
+
+    // 清理资源
+    ::SelectObject(hdcMemDC, hOldObj);
+    ::DeleteObject(hBitmap);
+    ::DeleteDC(hdcMemDC);
+    ::ReleaseDC(nullptr, hdcScreen);
+    ::ReleaseDC(hwnd, hdcWindow);
+
+    return true;
+}
+
 } //namespace ui
 
 #endif //DUILIB_BUILD_FOR_WIN
